macosx: Fix compile warnings in VLCLibraryAlbumTableCellView

+++ modules/gui/macosx/library/audio-library/VLCLibraryAlbumTableCellView.m
@@ -394,7 +394,7 @@ const CGFloat VLCLibraryAlbumTableCellViewDefaultHeight = 168.;
     NSArray * const tracks = album.mediaItems;
     const NSUInteger trackCount = tracks.count;
     const NSInteger clickedRow = _tracksTableView.clickedRow;
-    if (clickedRow >= 0 && clickedRow < trackCount) {
+    if (clickedRow >= 0 && (NSUInteger)clickedRow < trackCount) {
         VLCMediaLibraryMediaItem * const mediaItem = tracks[clickedRow];
         VLCLibraryRepresentedItem * const representedItem = [[VLCLibraryRepresentedItem alloc] initWithItem:mediaItem parentType:VLCMediaLibraryParentGroupTypeAlbum];
